Добавляет модульность в файлах js
